Add files via upload
The **Bionic Surgical Arm** project was designed to revolutionize the field of surgery by integrating cutting-edge computer vision technology to track the precise movements of a surgeon’s arm. The system is built to minimize human error during surgeries, ensuring higher accuracy and control. The core feature of the bionic arm is its ability to replicate a surgeon’s movements in real time, enabling remote surgeries to be performed in areas with limited access to medical facilities. By using advanced algorithms, the arm offers enhanced precision, reduces fatigue, and opens up the possibility of providing high-quality medical care in rural and underserved regions. This innovation not only improves surgical outcomes but also paves the way for greater accessibility and equality in healthcare delivery worldwide.